Move TRD RecoInputContainer to TRD DataFormats
